chore: Rename type directory to types
