+++ src/components/atoms/heading/index.tsx
@@ -1,7 +1,7 @@
 import type { FC } from "react";
 import tw, { css } from "twin.macro";
 import { useMemo } from "react";
-import type { HeadingProperties } from "./type/model";
+import type { HeadingProperties } from "./types/model";
 
 const HeadingBase = tw.h1`font-zen text-text text-2xl sm:text-5xl font-bold select-none`;
