+++ src/components/atoms/icon/index.tsx
@@ -1,7 +1,7 @@
 import { InlineIcon } from "@iconify/react";
 import { memo } from "react";
 import tw from "twin.macro";
-import type { IconProperties } from "./type/model";
+import type { IconProperties } from "./types/model";
 
 const IconBox = tw.div`flex items-center justify-center flex-shrink-0 w-6 h-6 sm:(w-12 h-12) min-h-full`;
