+++ src/components/atoms/icon/types/model.ts
