+++ src/components/organisms/about-section/types/model.ts
