+++ src/components/organisms/access-section/types/model.ts
