+++ src/components/organisms/footer/types/model.ts
