+++ src/components/organisms/sns-section/types/model.ts
